LocalNotificationManager, BdayTextView 리팩토링
